Started implimention
Added some implementation of functions. Not usable yet as you can't get to that page, but doing the back end so that I can wire up the front end to it when it's done.
Also made the database, not uploading the schema yet as I don't want that to be out there right now.