Small but impactful changes
Changed the ajax sent request to a POST from a GET, also made it more secure by not sending the password in the URL like a dumb dumb.
Added a new function to match to save the insult sent but only if it's a post request, else it will check if a match has been found, if not then match the player.